+++ group24/1148285693/learning2017/learning-basic/src/main/java/me/lzb/basic/InfixExpr.java
@@ -1,0 +1,19 @@
+package me.lzb.basic;
+
+/**
+ * Created by LZB on 2017/4/15.
+ */
+public class InfixExpr {
+
+    private String expr;
+
+    public InfixExpr(String expr) {
+        this.expr = expr;
+    }
+
+    public float evaluate() {
+
+        return 0.0f;
+    }
+
+}
