--- group24/1148285693/learning2017/learning-basic/src/main/java/me/lzb/datastructure/Iterator.java
@@ -1,10 +1,0 @@
-package me.lzb.datastructure;
-
-/**
- * Created by LZB on 2017/3/11.
- */
-public interface Iterator {
-	public boolean hasNext();
-	public Object next();
-
-}
