--- group24/1148285693/learning2017/learning-basic/src/main/java/me/lzb/datastructure/List.java
@@ -1,13 +1,0 @@
-package me.lzb.datastructure;
-
-/**
- * list接口
- * Created by LZB on 2017/3/11.
- */
-public interface List {
-	void add(Object o);
-	void add(int index, Object o);
-	Object get(int index);
-	Object remove(int index);
-	int size();
-}
