+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/attr/AttributeInfo.java
@@ -1,0 +1,22 @@
+package me.lzb.jvm.attr;
+
+/**
+ * Created by LZB on 2017/4/15.
+ */
+public abstract class AttributeInfo {
+
+    public static final String CODE = "Code";
+    public static final String CONST_VALUE = "ConstantValue";
+    public static final String EXCEPTIONS = "Exceptions";
+    public static final String LINE_NUM_TABLE = "LineNumberTable";
+    public static final String LOCAL_VAR_TABLE = "LocalVariableTable";
+    public static final String STACK_MAP_TABLE = "StackMapTable";
+
+    int attrNameIndex;
+    int attrLen;
+
+    public AttributeInfo(int attrNameIndex, int attrLen) {
+        this.attrNameIndex = attrNameIndex;
+        this.attrLen = attrLen;
+    }
+}
