+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/attr/LineNumberTable.java
@@ -1,0 +1,23 @@
+package me.lzb.jvm.attr;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Created by LZB on 2017/4/15.
+ */
+public class LineNumberTable extends AttributeInfo {
+    List<LineNumberItem> items = new ArrayList<>();
+
+
+    public void addLineNumberItem(LineNumberItem item) {
+        this.items.add(item);
+    }
+
+    public LineNumberTable(int attrNameIndex, int attrLen) {
+        super(attrNameIndex, attrLen);
+
+    }
+
+
+}
