+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/attr/LocalVariableTable.java
@@ -1,0 +1,20 @@
+package me.lzb.jvm.attr;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Created by LZB on 2017/4/15.
+ */
+public class LocalVariableTable extends AttributeInfo {
+    List<LocalVariableItem> items = new ArrayList<>();
+
+    public LocalVariableTable(int attrNameIndex, int attrLen) {
+        super(attrNameIndex, attrLen);
+    }
+
+    public void addLocalVariableItem(LocalVariableItem item) {
+        this.items.add(item);
+    }
+
+}
