+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/attr/StackMapTable.java
@@ -1,0 +1,15 @@
+package me.lzb.jvm.attr;
+
+/**
+ * Created by LZB on 2017/4/15.
+ */
+public class StackMapTable extends AttributeInfo {
+    private String originalCode;
+
+    public StackMapTable(int attrNameIndex, int attrLen, String code) {
+        super(attrNameIndex, attrLen);
+        this.originalCode = code;
+    }
+
+
+}
