+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/constant/ConstantInfo.java
@@ -5,6 +5,22 @@ package me.lzb.jvm.constant;
  */
 public abstract class ConstantInfo {
 
+    public static final String MAGIC_NUMBER = "cafebabe";
+
+    public static final int Class_info = 7;
+
+    public static final int Fieldref_info = 9;
+
+    public static final int Methodref_info = 10;
+
+    public static final int String_info = 8;
+
+    public static final int NameAndType_info = 12;
+
+    public static final int Utf8_info = 1;
+
+
+
     protected ConstantPool constantPool;
