--- group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/constant/JvmConstant.java
@@ -1,23 +1,0 @@
-package me.lzb.jvm.constant;
-
-/**
- * Jvm常量
- * Created by LZB on 2017/4/14.
- */
-public class JvmConstant {
-
-    public static final String MAGIC_NUMBER = "cafebabe";
-
-    public static final int Class_info = 7;
-
-    public static final int Fieldref_info = 9;
-
-    public static final int Methodref_info = 10;
-
-    public static final int String_info = 8;
-
-    public static final int NameAndType_info = 12;
-
-    public static final int Utf8_info = 1;
-
-}
