+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/constant/StringInfo.java
@@ -4,7 +4,7 @@ package me.lzb.jvm.constant;
  * Created by LZB on 2017/4/15.
  */
 public class StringInfo extends ConstantInfo {
-    private int type = JvmConstant.String_info;
+    private int type = ConstantInfo.String_info;
 
     private int index;
