+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/constant/UTF8Info.java
@@ -4,7 +4,7 @@ package me.lzb.jvm.constant;
  * Created by LZB on 2017/4/15.
  */
 public class UTF8Info extends ConstantInfo {
-    private int type = JvmConstant.Class_info;
+    private int type = ConstantInfo.Class_info;
 
     private String value;
