--- group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/exception/NotClassFileException.java
@@ -1,14 +1,0 @@
-package me.lzb.jvm.exception;
-
-/**
- * Created by LZB on 2017/4/14.
- */
-public class NotClassFileException extends Exception{
-    public NotClassFileException(String message, Exception cause) {
-        super(message, cause);
-    }
-
-    public NotClassFileException(String message) {
-        super(message);
-    }
-}
