+++ group24/1148285693/learning2017/mini-jvm/src/main/java/me/lzb/jvm/loader/ClassFileLoader.java
@@ -3,7 +3,6 @@ package me.lzb.jvm.loader;
 
 import me.lzb.common.utils.FileUtils;
 import me.lzb.jvm.clz.ClassFile;
-import me.lzb.jvm.exception.NotClassFileException;
 
 import java.io.File;
 import java.io.IOException;
@@ -12,51 +11,46 @@ import java.util.Iterator;
 import java.util.List;
 
 
-
 public class ClassFileLoader {
 
-	private List<String> clzPaths = new ArrayList<>();
+    private List<String> clzPaths = new ArrayList<>();
 
-	public byte[] readBinaryCode(String className) throws IOException {
+    public byte[] readBinaryCode(String className) throws IOException {
 //        String fileName = className.replaceAll(".*\\.", "")  + ".class";
 //        String pkg = className.replaceAll("\\.[^\\.]+$", "");
 //        String packagePath = pkg.replaceAll("\\.", "\\\\");
 
-        className = className.replace('.', File.separatorChar) +".class";
+        className = className.replace('.', File.separatorChar) + ".class";
         for (String s : clzPaths) {
             byte[] data = FileUtils.readByteCodes(s + className);
-            if(data != null){
+            if (data != null) {
                 return data;
             }
         }
 
         throw new IOException(className + "is not exist");
-	}
+    }
 
 
-	public void addClassPath(String path) {
+    public void addClassPath(String path) {
         clzPaths.add(path);
-	}
+    }
 
 
-	public String getClassPath(){
+    public String getClassPath() {
         StringBuilder buffer = new StringBuilder();
-        for (Iterator<String> iterator = clzPaths.iterator(); iterator.hasNext();) {
+        for (Iterator<String> iterator = clzPaths.iterator(); iterator.hasNext(); ) {
             buffer.append(iterator.next() + (iterator.hasNext() ? ";" : ""));
         }
         return buffer.toString();
-	}
+    }
 
 
-    public ClassFile loadClass(String className) throws IOException, NotClassFileException {
+    public ClassFile loadClass(String className) throws IOException {
         byte[] codes = this.readBinaryCode(className);
         ClassFileParser parser = new ClassFileParser(codes);
         return parser.parse();
     }
-
-
-
-
 
 
 }
