+++ group24/1148285693/learning2017/other/src/main/java/me/lzb/other/proxy/UserServiceImpl.java
@@ -1,0 +1,11 @@
+package me.lzb.other.proxy;
+
+/**
+ * Created by LZB on 2017/3/29.
+ */
+public class UserServiceImpl implements UserService {
+
+    public void add() {
+        System.out.println("--------------------add---------------");
+    }
+}
