+++ group24/1148285693/learning2017/other/src/test/java/me/lzb/other/proxy/ProxyTest.java
@@ -1,0 +1,25 @@
+package me.lzb.other.proxy;
+
+import org.junit.Test;
+
+/**
+ * Created by LZB on 2017/3/29.
+ */
+public class ProxyTest {
+
+    @Test
+    public void testProxy() throws Throwable {
+        // 实例化目标对象
+        UserService userService = new UserServiceImpl();
+
+        // 实例化InvocationHandler
+        MyInvocationHandler invocationHandler = new MyInvocationHandler(userService);
+
+        // 根据目标对象生成代理对象
+        UserService proxy = (UserService) invocationHandler.getProxy();
+
+        // 调用代理对象的方法
+        proxy.add();
+
+    }
+}
